Implement mechanism to lock and wipe memory
The generic Locked_Memory package can be instantiated with any definite
or indefinite type. It provides means to lock, wipe and unlock memory
regions allocated by associated element types.

Behind the scenes mlock(2), munlock(2) and memset(3) calls are used to
implement the functionality.

The Lock/Wipe/Unlock operations of the generic expect an access to the
element type. This ensures that all objects (scalars as well) are passed
by reference.