Add initial implementation of cert chain validation
Currently the code checks that user certificates are valid X509
certificates and verifies that their validity period is ok.

When adding a certificate to the chain the same checks as for the user
certificate are performed. Additionally the signature of the last
certificate in the chain is verified with the public key of the new
cert. It is also checked that the new certificate is a CA certificate.
If all these checks are ok the new certificate is made the new root of
the trustchain.

The handling of the validity period of the cc context is currently not
implemented. These values should be adjusted according to the min/max
values of the certificates in the trustchain.