Add XML config transformation tool
The config tool can be used to create TKM and IKE configurations from a source
file in XML format.
The usage of the cfg_tool binary is as follows:

Usage: obj/cfg_tool -c <config> -i <ikefile> -t <tkmfile> -s <schema>
  -c config XML file
  -i filename of IKE script to be written
  -t filename of TKM config to be written
  -s XML schema file (optional, default: schema/tkmconfig.xsd)
  -v version information

The source config file and either a filename for the IKE script or the TKM
config must be provided.